fix: update docs and examples for moved LangChain providers
Update HuggingFace pipeline config examples and streaming docs
to reference new import paths under integrations/langchain/.
