+++ docs/run-rails/using-python-apis/streaming.md
@@ -218,11 +218,11 @@ For streaming while using the Guardrails API server, refer to [](../using-fastap
 We also support streaming for LLMs deployed using `HuggingFacePipeline`.
 One example is provided in the [HF Pipeline Dolly](https://github.com/NVIDIA-NeMo/Guardrails/tree/develop/examples/configs/llm/hf_pipeline_dolly/README.md) configuration.
 
-To use streaming for HF Pipeline LLMs, you need to create an `nemoguardrails.llm.providers.huggingface.AsyncTextIteratorStreamer` streamer object,
+To use streaming for HF Pipeline LLMs, you need to create an `nemoguardrails.integrations.langchain.providers.huggingface.AsyncTextIteratorStreamer` streamer object,
 add it to the `kwargs` of the pipeline and to the `model_kwargs` of the `HuggingFacePipelineCompatible` object.
 
 ```python
-from nemoguardrails.llm.providers.huggingface import AsyncTextIteratorStreamer
+from nemoguardrails.integrations.langchain.providers.huggingface import AsyncTextIteratorStreamer
 
 # instantiate tokenizer object required by LLM
 streamer = AsyncTextIteratorStreamer(tokenizer, skip_prompt=True)
