+++ examples/configs/llm/hf_pipeline_dolly/config.py
@@ -14,10 +14,10 @@
 # limitations under the License.
 from functools import lru_cache
 
-from nemoguardrails.llm.helpers import get_llm_instance_wrapper
-from nemoguardrails.llm.providers.huggingface import HuggingFacePipelineCompatible
 from transformers import AutoConfig, AutoModelForCausalLM, AutoTokenizer, pipeline
 
+from nemoguardrails.integrations.langchain.helpers import get_llm_instance_wrapper
+from nemoguardrails.integrations.langchain.providers.huggingface import HuggingFacePipelineCompatible
 from nemoguardrails.llm.providers import register_llm_provider
 
 
@@ -40,7 +40,7 @@ def get_dolly_v2_3b_llm(streaming: bool = True):
 
     # If we want streaming, we create a streamer.
     if streaming:
-        from nemoguardrails.llm.providers.huggingface import AsyncTextIteratorStreamer
+        from nemoguardrails.integrations.langchain.providers.huggingface import AsyncTextIteratorStreamer
 
         streamer = AsyncTextIteratorStreamer(tokenizer, skip_prompt=True)
         params["streamer"] = streamer
